+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 
+## 1.0.3 - 2020-04-18
+  - Improved customer interface
+  - Added undocumented Razorpay customer's address api to the SDK.
+
+## 1.0.2 - 2020-04-17
+  - Added git repo to the package.json
+  - Added `index.d.ts` to the root of the package.
+  
+## 1.0.1 - 2020-04-17
+  - [Bug] Resolved internal target issue
+
 ## 1.0.0 - 2020-04-16
   - Initial Release
   - [Docs] Add readme.md
   - [Docs] Add Changelog
-  - 
-## 1.0.1 - 2020-04-17
-  - [Bug] Resolved internal target issue
+